# docs/swagger.yaml
basePath: /
definitions:
  controller.Hewan:
    properties:
      _id:
        example: "123456789"
        type: string
      jenis:
        example: Anjing
        type: string
      ras:
        example: Pitbull
        type: string
      umur:
        example: "8"
        type: string
    type: object
  controller.MakananHewan:
    properties:
      _id:
        example: "123456789"
        type: string
      bahan:
        example: Daging Ikan
        type: string
      berat:
        example: "4"
        type: string
      harga:
        example: "500000"
        type: string
      hewan:
        $ref: '#/definitions/controller.Hewan'
      jenismakanan:
        example: Basah
        type: string
      merk:
        example: Whiskas
        type: string
      rasa:
        example: Tuna
        type: string
      tanggal:
        example: "2024-09-01T00:00:00Z"
        format: date-time
        type: string
    type: object
  controller.ReqHewan:
    properties:
      jenis:
        example: Anjing
        type: string
      ras:
        example: Pitbull
        type: string
      umur:
        example: "8"
        type: string
    type: object
  controller.ReqMakananHewan:
    properties:
      bahan:
        example: Daging Ikan
        type: string
      berat:
        example: "4"
        type: string
      harga:
        example: "500000"
        type: string
      hewan:
        $ref: '#/definitions/controller.ReqHewan'
      jenismakanan:
        example: Basah
        type: string
      merk:
        example: Whiskas
        type: string
      rasa:
        example: Tuna
        type: string
      tanggal:
        example: "2024-09-01T00:00:00Z"
        format: date-time
        type: string
    type: object
host: backend-tb-97f8facdbf4c.herokuapp.com
info:
  contact:
    email: 714220048@std.ulbi.ac.id
    name: API Support
    url: https://github.com/alfianbr16
  description: This is a sample swagger for Fiber
  title: TES SWAGGER ULBI
  version: "1.0"
paths:
  /delete/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus data makanan.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete Data MakananHewan.
      tags:
      - Makanan
  /ins:
    post:
      consumes:
      - application/json
      description: Input data makanan.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqMakananHewan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.MakananHewan'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert Data MakananHewan.
      tags:
      - Makanan
  /makanan:
    get:
      consumes:
      - application/json
      description: Mengambil semua data makanan.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.MakananHewan'
      summary: Get All Data MakananHewan.
      tags:
      - Makanan
  /makanan/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data makanan.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.MakananHewan'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data MakananHewan.
      tags:
      - Makanan
  /update/{id}:
    put:
      consumes:
      - application/json
      description: Ubah data makanan.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqMakananHewan'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.MakananHewan'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update Data MakananHewan.
      tags:
      - Makanan
schemes:
- https
- http
swagger: "2.0"